\name{dcorVS-package}
\alias{dcorVS-package}
\docType{package}
\title{
Variable Selection Algorithms Using the Distance Correlation.
}
\description{
The 'FBED' and 'mmpc' variable selection algorithms have been implemented using the distance correlation.
}
\details{
\tabular{lll}{
Package: \tab dcorVS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2023-10-17\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Szekely G.J., Rizzo M.L. and Bakirov N.K. (2007). Measuring and Testing Independence
by Correlation of Distances. Annals of Statistics, 35(6): 2769--2794.

Szekely G.J. and Rizzo M. L. (2014). Partial distance correlation with methods for dissimilarities.
Annals of Statistics, 42(6): 2382--2412.

Huo X. and Szekely G.J. (2016). Fast computing for distance covariance.
Technometrics, 58(4): 435--447.

Tsamardinos I., Aliferis C. F. and Statnikov A. (2003). Time and sample efficient discovery of Markov
blankets and direct causal relations. In Proceedings of the ninth ACM SIGKDD international Conference
on Knowledge Discovery and Data Mining (pp. 673--678). ACM.

Brown L. E., Tsamardinos I. and Aliferis C. F. (2004). A novel algorithm for scalable and accurate
Bayesian network learning. Medinfo, 711--715.

Borboudakis G. and Tsamardinos I. (2019). Forward-backward selection with early dropping.
Journal of Machine Learning Research, 20(8): 1--39.
}



